sweeper/requirements.txt
solana==0.34.0
solders==0.21.0
requests==2.31.0
python-dotenv==1.0.1
